SUPPORT-91253: updated backward compatibility
